third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-convolvernode-interface/active-processing.https-expected.txt

This is a testharness.js-based test.
PASS # AUDIT TASK RUNNER STARTED.
PASS > [Test] Active processing for ConvolverNode
PASS   Construction of graph with convolver did not throw an exception.
FAIL X Number of output channels of convolver expected to have the value sequence of [1,2,1] but got 2 at index 0. assert_true: expected true got false
FAIL < [Test] 1 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 1 tasks were failed. assert_true: expected true got false
Harness: the test ran to completion.
